Obessvations were fixed according feedback.
