--- src/main/java/com/ecossio7/App.java
@@ -1,13 +1,0 @@
-package com.ecossio7;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
