+++ src/test/java/DiagramClassInit.png
