+++ src/test/java/anotations/Regression.java
@@ -1,11 +1,11 @@
-package Anotations;
+package anotations;
 
 import org.junit.jupiter.api.Tag;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-@Tag("smoke")
+@Tag("regression")
 @Retention(value = RetentionPolicy.RUNTIME)
-public @interface Smoke {
+public @interface Regression {
 }
