+++ src/test/java/com/ecossio7/participants/ParticipantsEmptyTests.java
@@ -1,8 +1,8 @@
 package com.ecossio7.participants;
 
-import Anotations.Regression;
+import anotations.Regression;
 import com.microsoft.playwright.APIRequestContext;
-import models.RError;
+import models.Error;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,33 +24,30 @@ public class ParticipantsEmptyTests extends BaseTest {
     @Test
     @Regression
     void createParticipantTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = participantsRequests.create(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 
     @Test
     @Regression
     void updateParticipantTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = participantsRequests.update(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 
     @Test
     @Regression
     void partialParticipantTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = participantsRequests.partialUpdate(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 }
