+++ src/test/java/com/ecossio7/videojuegos/VideoJuegosEmptyTests.java
@@ -1,8 +1,8 @@
 package com.ecossio7.videojuegos;
 
-import Anotations.Regression;
+import anotations.Regression;
 import com.microsoft.playwright.APIRequestContext;
-import models.RError;
+import models.Error;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,33 +23,30 @@ public class VideoJuegosEmptyTests extends BaseTest {
     @Test
     @Regression
     void createVideoJuegoTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = videoJuegoRequests.create(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 
     @Test
     @Regression
     void updateVideoJuegoTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = videoJuegoRequests.update(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 
     @Test
     @Regression
     void partialVideoJuegoTest() {
-        final var body = "{}";
-        requestOptions.setData(body);
+        requestOptions.setData("{}");
         apiResponse = videoJuegoRequests.partialUpdate(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(400, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("Debe especificar un payload", actualError.mensaje());
     }
 }
