+++ src/test/java/com/ecossio7/videojuegos/VideoJuegosNoAuthTests.java
@@ -1,10 +1,10 @@
 package com.ecossio7.videojuegos;
 
-import Anotations.Regression;
+import anotations.Regression;
 import com.google.gson.JsonParser;
 import com.microsoft.playwright.APIRequestContext;
-import models.RError;
-import models.RVideoJuego;
+import models.Error;
+import models.VideoJuego;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -27,8 +27,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
     @Regression
     void getAllVideoJuegosTest() {
         apiResponse = videoJuegoRequests.getAll(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -37,8 +37,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
     void findVideoJuegoTest() {
         requestOptions.setQueryParam("nombre", "Pac-Man");
         apiResponse = videoJuegoRequests.getAll(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -48,8 +48,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
         requestOptions.setQueryParam("sortBy", "epoca");
         requestOptions.setQueryParam("order", "asc");
         apiResponse = videoJuegoRequests.getAll(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -59,8 +59,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
         requestOptions.setQueryParam("filterBy", "genero");
         requestOptions.setQueryParam("value", "comedia");
         apiResponse = videoJuegoRequests.getAll(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -68,22 +68,22 @@ public class VideoJuegosNoAuthTests extends BaseTest {
     @Regression
     void getVideoJuegoTest() {
         apiResponse = videoJuegoRequests.getById(25, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
     @Test
     @Regression
     void createVideoJuegoTest() {
-        final var participant = RVideoJuego.generateVideoJuego();
+        final var participant = VideoJuego.generateVideoJuego();
         final var jsonString = gson.toJson(participant);
         final var jsonObject = JsonParser.parseString(jsonString).getAsJsonObject();
         jsonObject.remove(("id"));
         requestOptions.setData(jsonObject);
         apiResponse = videoJuegoRequests.create(requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -93,8 +93,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
         final var body = Files.readAllBytes(Paths.get("src/test/resources/videoJuego.json"));
         requestOptions.setData(body);
         apiResponse = videoJuegoRequests.update(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -112,8 +112,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
                 """;
         requestOptions.setData(body);
         apiResponse = videoJuegoRequests.partialUpdate(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 
@@ -121,8 +121,8 @@ public class VideoJuegosNoAuthTests extends BaseTest {
     @Regression
     void deleteVideoJuegoTest() {
         apiResponse = videoJuegoRequests.delete(5, requestOptions);
-        final var actualError = gson.fromJson(apiResponse.text(), RError.class);
         Assertions.assertEquals(401, apiResponse.status());
+        final var actualError = gson.fromJson(apiResponse.text(), Error.class);
         Assertions.assertEquals("No autorizado", actualError.mensaje());
     }
 }
