+++ src/test/java/models/Error.java
@@ -1,0 +1,4 @@
+package models;
+
+public record Error(String mensaje) {
+}
