--- src/test/java/models/RError.java
@@ -1,4 +1,0 @@
-package models;
-
-public record RError(String mensaje) {
-}
