+++ src/test/java/models/Reaccion.java
@@ -1,10 +1,10 @@
 package models;
 
-public record RReaccion(int likes, int dislikes) {
-    public static RReaccion generateRReaccion() {
+public record Reaccion(int likes, int dislikes) {
+    public static Reaccion generateRReaccion() {
         final var faker = new net.datafaker.Faker();
         final var likes = faker.number().numberBetween(0, 1000);
         final var dislikes = faker.number().numberBetween(0, 1000);
-        return new RReaccion(likes, dislikes);
+        return new Reaccion(likes, dislikes);
     }
 }
